Rename Entity module in tellduslive
