+++ homeassistant/components/tellduslive/entity.py
